backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai>=0.3.0
transformers>=4.30.0
torch>=2.6.0
gunicorn==21.2.0
pydantic==2.5.2
langdetect==1.0.9
numpy>=1.21.0
pandas>=2.0.0
PyPDF2==3.0.1
pdfplumber>=0.9.0
sumy==0.11.0
nltk==3.8.1
werkzeug==3.0.1
